# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ruijie API Proxy
  version: 1.0.0
  description: |
    Architecturally robust Express proxy API for Ruijie Cloud upstream services.

    ### Core Principles
    - **Clean Architecture**: Adheres to modular design with separated concerns.
    - **Consistent Response Envelopes**: Every API endpoint (excluding Swagger UI) returns a standardized JSON structure.
    - **Semantic Error Handling**: Specific error classes and codes for predictable client-side handling.
    - **Data Normalization**: Raw upstream data is formatted for human and AI readability (Mbps, HH:mm, standard timestamps).

    ### Response Envelope Contract
    - **Success**: `{ "success": true, "message": "OK", "data": { ... }, "meta": { ... } }`
    - **Error**: `{ "success": false, "message": "...", "error": { "httpStatus": 400, "name": "ValidationError", "details": { ... } } }`

    ### Special Error Mapping
    If a network group is not synchronized in the Ruijie Cloud console, the proxy returns:
    - `409 Conflict`
    - `error.code: "USERGROUP_NOT_SYNCED"`
    - Client Action: Prompt user to refresh network group list.

servers:
  - url: http://localhost:3000
    description: Local Development Server

security:
  - bearerAuth: []

paths:
  /login:
    post:
      tags: [Authentication]
      summary: Authenticate with Ruijie app credentials
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponseEnvelope"

  /login/vip:
    post:
      tags: [Authentication]
      summary: Authenticate via VIP credential mapping
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VipLoginRequest"
      responses:
        "200":
          description: VIP authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponseEnvelope"

  /auth/core/projects:
    get:
      tags: [Authentication]
      summary: List synchronized projects for current session
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"

  /auth/core/tenant:
    get:
      tags: [Authentication]
      summary: Get current tenant details
      responses:
        "200":
          description: Tenant information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"

  /network_group:
    get:
      tags: [Discovery]
      summary: List all leaf network groups
      responses:
        "200":
          description: Array of network groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkGroupListEnvelope"

  /packages:
    get:
      tags: [Packages]
      summary: List all packages (user groups) for a group
      parameters:
        - name: groupId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageListEnvelope"

  /packages/create:
    post:
      tags: [Packages]
      summary: Create a new package
      description: Orchestrates the creation of both an Auth Profile and a User Group.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Package created successfully

  /packages/{groupId}:
    post:
      tags: [Packages]
      summary: Update package group settings
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Package updated successfully

  /vouchers/active:
    get:
      tags: [Vouchers]
      summary: List active (in-use) vouchers
      parameters: &VoucherParams
        - name: groupId
          in: query
          required: true
          schema:
            type: string
        - name: start
          in: query
          schema:
            type: integer
            default: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of active vouchers grouped by date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherGroupListEnvelope"

  /vouchers/remain:
    get:
      tags: [Vouchers]
      summary: List remaining (unused) vouchers
      parameters: *VoucherParams
      responses:
        "200":
          description: List of unused vouchers grouped by date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherGroupListEnvelope"

  /vouchers/expired:
    get:
      tags: [Vouchers]
      summary: List expired vouchers
      parameters: *VoucherParams
      responses:
        "200":
          description: List of expired vouchers grouped by date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherGroupListEnvelope"
    delete:
      tags: [Vouchers]
      summary: Batch delete all expired vouchers for a group
      parameters:
        - name: groupId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Expired vouchers deleted

  /vouchers/status:
    get:
      tags: [Vouchers]
      summary: Get voucher count summary
      parameters:
        - name: groupId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Summary counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherStatusEnvelope"

  /vouchers/performance:
    get:
      tags: [Vouchers]
      summary: Get voucher performance metrics
      parameters:
        - name: groupId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Financial metrics for vouchers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherPerformanceEnvelope"

  /vouchers/generate:
    post:
      tags: [Vouchers]
      summary: Generate new vouchers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoucherGenerateRequest"
      responses:
        "200":
          description: Vouchers generated successfully

  /clients:
    get:
      tags: [Clients]
      summary: List authenticated clients
      parameters:
        - name: group_id
          in: query
          required: true
          schema:
            type: string
        - name: page_index
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of authenticated clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListEnvelope"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: appid::token

  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: OK }
        data: { type: object }
        meta: { type: object }

    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        error:
          type: object
          properties:
            httpStatus: { type: integer }
            name: { type: string }
            code: { type: string }
            details: { type: object, nullable: true }

    LoginRequest:
      type: object
      required: [appid, secret]
      properties:
        appid: { type: string }
        secret: { type: string }

    VipLoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }

    LoginResponseEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - properties:
            data:
              type: object
              properties:
                authorization:
                  { type: string, description: "Composite Bearer token" }
                appid: { type: string }
                tenantName: { type: string }

    NetworkGroupListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - properties:
            data:
              type: array
              items:
                type: object
                properties:
                  name: { type: string }
                  groupId: { type: string }

    PackageListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - properties:
            data:
              type: array
              items:
                type: object

    VoucherGroupListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - properties:
            data:
              type: array
              items:
                type: object
                properties:
                  date: { type: string }
                  timeType: { type: string }
                  vouchers: { type: array, items: { type: object } }

    VoucherStatusEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - properties:
            data:
              type: object
              properties:
                total: { type: integer }
                used: { type: integer }
                expired: { type: integer }

    VoucherPerformanceEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - properties:
            data:
              type: object
              properties:
                lastDay:
                  type: object
                  properties:
                    count: { type: integer }
                    price: { type: number }
                monthly:
                  type: object
                  properties:
                    count: { type: integer }
                    price: { type: number }

    VoucherGenerateRequest:
      type: object
      required: [groupId, userGroupId, count]
      properties:
        groupId: { type: string }
        userGroupId: { type: integer }
        profile: { type: string, description: "Alias for authProfileId" }
        count: { type: integer }

    ClientListEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - properties:
            data:
              type: object
              properties:
                list:
                  type: array
                  items:
                    type: object
                    properties:
                      mac: { type: string }
                      staModel: { type: string }
                      ip: { type: string }
                      duration: { type: string, example: "01:45" }